made tabs in insert functional
